+++ src/server/rpc/procedures/met-museum/seed.js
@@ -1,6 +1,6 @@
 #!/bin/env node
 
-const seeder = require('../utils/csv-toolset');
+const seeder = require('../utils/csv-loader');
 const MetMuseumCol = require('./database');
 
 seeder(MetMuseumCol, {url: 'https://media.githubusercontent.com/media/metmuseum/openaccess/master/MetObjects.csv'});
